Refactor shortcodeFactory to provide page to paired shortcodes, use chalk factory to log warnings in section.js
