+++ packages/11ty/_plugins/components/index.js
@@ -1,5 +1,5 @@
 const addComponentTag = require('./addComponentTag')
-const shortcodeFactory = require('./addShortcode')
+const shortcodeFactory = require('./shortcodeFactory')
 
 // Shortcode components
 const components = require('../../_includes/components')
@@ -11,7 +11,7 @@ const components = require('../../_includes/components')
  * @param      {Object}  options         options
  */
 module.exports = function(eleventyConfig, collections, options) {
-  const addShortcode = shortcodeFactory(eleventyConfig, collections)
+  const { addShortcode } = shortcodeFactory(eleventyConfig, collections)
 
   for (const component in components) {
     switch(component) {
