+++ packages/11ty/_plugins/shortcodes/index.js
@@ -8,20 +8,16 @@ const figure = require('./figure')
 const figureGroup = require('./figureGroup')
 const ref = require('./figureRef')
 const section = require('./section.js')
-const shortcodeFactory = require('../../_plugins/components/addShortcode')
+const shortcodeFactory = require('../components/shortcodeFactory')
 const title = require('./title')
 const tombstone = require('./tombstone')
 
 module.exports = function(eleventyConfig, collections, options) {
-  const addShortcode = shortcodeFactory(eleventyConfig, collections)
+  const { addShortcode, addPairedShortcode } = shortcodeFactory(eleventyConfig, collections)
 
   addComponentTag(eleventyConfig, 'annoref', annoref)
-  eleventyConfig.addPairedShortcode('backmatter', function(content, ...args) {
-    return backmatter(eleventyConfig)(content, ...args)
-  })
-  eleventyConfig.addPairedShortcode('section', function(content, ...args) {
-    return section(eleventyConfig)(content, ...args)
-  })
+  addPairedShortcode('backmatter', backmatter)
+  addPairedShortcode('section', section)
   addShortcode('bibliography', bibliography)
   addShortcode('cite', cite)
   addComponentTag(eleventyConfig, 'contributors', contributors)
